#33 wip split api code
